backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
pymongo==4.3.3
pandas==1.5.3
scikit-learn==1.2.2
torch==2.0.0
numpy==1.24.2
gensim==4.3.1
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==20.1.0
pytest==7.3.1 